requirements.txt
anthropic>=0.8.0
arxiv>=2.0.0
openai>=1.12.0
psycopg2-binary>=2.9.0
sqlite3
python-dotenv>=1.0.0
requests>=2.31.0
torch>=2.2.0
transformers>=4.37.0
typing>=3.7.4
